[REFACTOR] #107 unify userStores To AuthStore.js
